text_analysis_service/requirements.txt
flask
boto3
python-dotenv
openai
PyPDF2